zink: add arrayness I/O matching
When an XFB-enabled pre-raster stage is combined with a fragment
shader that addresses the array that contains the XFB variables
indirectly, there's a mismatch of arrayness that occurs between
these two stages.

Overcome this hurdle by lowering the indirectly addressed FS
array to be directly accessed through a local copy of the array.

Part-of: <https://gitlab.freedesktop.org/mesa/mesa/-/merge_requests/37602>